functions/requirements.txt
firebase-functions
firebase-admin
gradio_client
python-dotenv